refactor: js->jsx 파일로 변경
팀 컨벤션에 맞게 jsx파일로 변경
