+++ client/src/components/dropDown/index.js
@@ -1,1 +1,1 @@
-export { default } from './dropDown';
+export { default } from './dropDown.jsx';
