+++ client/src/components/header/index.js
@@ -1,1 +1,1 @@
-export { default } from './header';
+export { default } from './header.jsx';
