+++ client/src/components/sideMenu/commonSidePart/index.js
@@ -1,1 +1,1 @@
-export { default } from './commonSidePart';
+export { default } from './commonSidePart.jsx';
