+++ client/src/components/sideMenu/labelPart/index.js
@@ -1,1 +1,1 @@
-export { default } from './labelPart';
+export { default } from './labelPart.jsx';
