+++ client/src/pages/IssueDetail/mainPart/index.js
@@ -1,1 +1,1 @@
-export { default } from './issueDetailPart';
+export { default } from './issueDetailPart.jsx';
