+++ client/src/pages/IssueDetail/sidePart/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import SideAssigneePart from './sideAssigneePart';
-import SideLabelPart from './sideLabelPart';
-import SideMilestonePart from './sideMilestonePart';
+import SideAssigneePart from './sideAssigneePart.jsx';
+import SideLabelPart from './sideLabelPart.jsx';
+import SideMilestonePart from './sideMilestonePart.jsx';
 
 const SidePartStyle = styled.div`
   position: fixed;
